Update newfile.html
Ok so @bamanina you'll notice that I moved lines 44 - 47 so that the indentation was aligned better.  This is because of the editor.  Some editors add tabs / spaces or additional unnecessary white space to your code.  The goal is to have everything lined up and indented properly so that it's easy to read. This will be very important when working on files that are white space sensitive like haml.  :)